+++ src/main/java/com/hta2405/unite/action/admin/AdminHolidayInsertAction.java
@@ -1,5 +1,7 @@
-package com.hta2405.unite.action;
+package com.hta2405.unite.action.admin;
 
+import com.hta2405.unite.action.Action;
+import com.hta2405.unite.action.ActionForward;
 import com.hta2405.unite.dao.HolidayDao;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
@@ -10,15 +12,16 @@ import java.time.LocalDate;
 
 import static com.hta2405.unite.util.CommonUtil.alertAndGoBack;
 
-public class AdminHolidayDeleteAction implements Action {
+public class AdminHolidayInsertAction implements Action {
     @Override
     public ActionForward execute(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
         LocalDate date = LocalDate.parse(req.getParameter("date"));
-        if (new HolidayDao().getHolidayName(date) == null) {
-            return alertAndGoBack(resp, "휴일 삭제 실패, 휴일이 아닙니다.");
+        String HolidayName = req.getParameter("holidayName");
+        if (new HolidayDao().getHolidayName(date) != null) {
+            return alertAndGoBack(resp, "휴일 등록 실패, 이미 등록된 휴일입니다.");
         }
-        if (new HolidayDao().deleteHoliday(date) != 1) {
-            return alertAndGoBack(resp, "휴일 삭제 실패");
+        if (new HolidayDao().insertHoliday(date, HolidayName) != 1) {
+            return alertAndGoBack(resp, "휴일 등록 실패");
         }
         return new ActionForward(true, req.getContextPath() + "/admin/holiday");
     }
