+++ src/main/java/com/hta2405/unite/action/admin/AdminHolidayWeekendAction.java
@@ -1,5 +1,7 @@
-package com.hta2405.unite.action;
+package com.hta2405.unite.action.admin;
 
+import com.hta2405.unite.action.Action;
+import com.hta2405.unite.action.ActionForward;
 import com.hta2405.unite.dao.HolidayDao;
 import com.hta2405.unite.util.CommonUtil;
 import jakarta.servlet.ServletException;
