+++ src/main/java/com/hta2405/unite/action/attend/AttendEmpDetailAction.java
@@ -1,5 +1,7 @@
-package com.hta2405.unite.action;
+package com.hta2405.unite.action.attend;
 
+import com.hta2405.unite.action.Action;
+import com.hta2405.unite.action.ActionForward;
 import com.hta2405.unite.dao.EmpDao;
 import com.hta2405.unite.dto.Emp;
 import com.hta2405.unite.util.AttendUtil;
