+++ src/main/java/com/hta2405/unite/action/attend/AttendInAction.java
@@ -1,6 +1,8 @@
-package com.hta2405.unite.action;
+package com.hta2405.unite.action.attend;
 
 import com.google.gson.JsonObject;
+import com.hta2405.unite.action.Action;
+import com.hta2405.unite.action.ActionForward;
 import com.hta2405.unite.dao.AttendDao;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
