+++ src/main/java/com/hta2405/unite/action/attend/AttendInfoAction.java
@@ -1,6 +1,8 @@
-package com.hta2405.unite.action;
+package com.hta2405.unite.action.attend;
 
 import com.google.gson.JsonObject;
+import com.hta2405.unite.action.Action;
+import com.hta2405.unite.action.ActionForward;
 import com.hta2405.unite.dao.AttendDao;
 import com.hta2405.unite.dto.Attend;
 import jakarta.servlet.ServletException;
