+++ src/main/java/com/hta2405/unite/action/attend/AttendVacationEmpListAction.java
@@ -1,14 +1,16 @@
-package com.hta2405.unite.action;
+package com.hta2405.unite.action.attend;
 
+import com.hta2405.unite.action.Action;
+import com.hta2405.unite.action.ActionForward;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 
 import java.io.IOException;
 
-public class MyPagePasswordChangeAction implements Action {
+public class AttendVacationEmpListAction implements Action {
     @Override
     public ActionForward execute(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        return new ActionForward(false, "/WEB-INF/views/mypage/password.jsp");
+        return null;
     }
 }
