+++ src/main/java/com/hta2405/unite/action/mypage/MyPagePasswordChangeAction.java
@@ -1,14 +1,16 @@
-package com.hta2405.unite.action;
+package com.hta2405.unite.action.mypage;
 
+import com.hta2405.unite.action.Action;
+import com.hta2405.unite.action.ActionForward;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 
 import java.io.IOException;
 
-public class AttendVacationEmpListAction implements Action {
+public class MyPagePasswordChangeAction implements Action {
     @Override
     public ActionForward execute(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        return null;
+        return new ActionForward(false, "/WEB-INF/views/mypage/password.jsp");
     }
 }
