+++ src/main/java/com/hta2405/unite/controller/AttendFrontController.java
@@ -1,6 +1,7 @@
 package com.hta2405.unite.controller;
 
 import com.hta2405.unite.action.*;
+import com.hta2405.unite.action.attend.*;
 import jakarta.servlet.RequestDispatcher;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.annotation.WebServlet;
